Move sources for `uprobestats-test` up one level
`packages/modules/UprobeStats/src` was crowded with code for a few
different things. Everything in the `test` directory is concerned with
integration tests. Move that directory up one level so it is not
conflated with the other things in `src` that are unrelated to said
integration tests.

Bug: 408256298
Test: TH
Flag: EXEMPT TEST_ONLY
Change-Id: Idd14b7d90a768506e48363434c032d736595b44f