Rename and refactor `uprobestats-test` sources
All of the java files for `uprobestats-test` were previously named
"SmokeTest" something. This was a relic of the original test, which was
intended only to make sure that we could successfully attach eBPF
programs and write to statsd.

Rename to "UprobeStatsTest" since this is now a full blown integration
test suite and not just a simple "smoke test".

Additionally, make the package `com.android.uprobestats` to match the
apex and package naming used elsewhere.

Finally, move the common setup code to a JUnit TestRule for easier
reuse.

Bug: 408256298
Test: TH
Flag: EXEMPT TEST_ONLY
Change-Id: I5ddf5f626c080cac02fd30d176813e3d712b5dcf