Implement UprobeStatsService
This change introduces the uprobestats service implementation, but does
not (yet) start the service on device. A subsequent change will
introduce:
- registering the service in the main `uprobestats` executable
- calling the service from the client library
- refactoring integration tests given the service

It also refactors the `bpf_perf_event_open` FFI function to return an
`OwnedFd` to rust. This allows the file descriptor associated with a
given bpf program to be closed when a task completes, as opposed to when
the uprobestats executable itself completes (which will not always
happen right away in the binder service scenario).

Bug: 436244914
Test: TH
Flag: build.RELEASE_UPROBESTATS_SERVICE
Change-Id: I83d8fe599b2ddf7199ee5a4b79914b8584f9a572