Use SYS_ADMIN instead of PERFMON capability
In Linux 6.1, the PERFMON capability is sufficient to attach uprobes via
the perf_event_open syscall [1]. However, later kernel versions (6.16
upstream, but even earlier in the Android tree) require SYS_ADMIN for
the same [2].

Replace the PERFMON capability with SYS_ADMIN, which, for the sake of
this check is a superset of the former [3]. This will allow the system
to function on both newer and older kernels.

Furthermore, enable the basic integration test for uprobestats that was
disabled for x86 (cuttlefish) targets. The problem wasn't x86, it was
the (newer) kernel version used on those test targets.

[1]:
https://github.com/torvalds/linux/blob/v6.1/kernel/events/core.c#L10095-L10096
[2]:
https://github.com/torvalds/linux/commit/ba677dbe77af5ffe6204e0f3f547f3ba059c6302
[3]:
https://github.com/torvalds/linux/blob/v6.17/include/linux/capability.h#L190-L193

Bug: 455573923
Test: TH
Test: atest uprobestats-test
Flag: EXEMPT BUGFIX
Change-Id: Ifd76b257ac544623e97c85ff19ae87eca8f96873