Replace `hello_uprobestats` with `uprobestats-manual-test`
Recent refactoring to run uprobestats as a binder service broke the
ability to simply drop a file in a certain location and then start
uprobestats with `ctl.start`.

Replace `hello_uprobestats.py` with a test case that will have the same
effect - starting uprobestats with a given configuration. This can be
used for manual testing.

Bug: N/A
Test: UPROBESTATS_TEST_CONFIG=binder atest uprobestats-manual-test
Flag: EXEMPT HOST_MODE
Change-Id: Ib388b639daf5d11fa594121505115c74fd0ec60f