Downgrade a bunch of log statements to TRACE
The DEBUG level was being used for most verbose logging, making the logs
very chatty (even for local debugging).

Downgrade many statements to TRACE so that we can set the logging level
to DEBUG when we want to see some logs for local debugging, but not
necessarily every single detail of uprobestats' execution.

Bug: 408256298
Test: TH
Flag: EXEMPT PURE_REFACTOR
Change-Id: I436b8c970eabf773893cf65dd70a2908d41d8d15