Bind to DynamicInstrumentationEventConsumer as system_server
An incoming fix will cause the bind to throw an exception, as the
consumer needs to be bound to by System Server, not the caller of
UprobeStatsBridgeService (which will be uprobestats).

Additionally, fix flakiness from slow ITestService. The test awaits the
server has sent all the events, but not that the client has received and
processed them.

Bug: 457625065
Bug: 460344584
Test: atest CtsUprobeStatsTest --iterations 10
Flag: EXEMPT BUGFIX
Change-Id: Id4193fd05dec8d7a33fea0cfb3a057f66416fec0