Abstraction for writing statsd atoms
To move more code into the uprobestats_core crate, a facade over statsd
is required, since statsd doesn't exist on the host.

This allows BPF handlers to be implemented and tested in the core
library. This also enables test code to use mocked implementation(s) and
easily assert that logic that determines when to write atoms is correct.

Additionally:
- create a host supported library for
uprobestats_bpf_structs, and modify the source file to use e.g. uint32_t
instead of __u32.
- update AStatsSocket to return `Result` on `write`, since it is a fallible operation.

Bug: 413078491
Test: TH
Flag: EXEMPT PURE_REFACTOR
Change-Id: I245679eb5a591bbed47f2f5f7fd75c6f3ee3ee80